Refactored file layout and clean up
