+++ scalalib/src/mill/scalalib/buildfile/MillBuildModule.scala
@@ -1,96 +1,13 @@
 package mill.scalalib.buildfile
 
-import mill.api.JsonFormatters._
-import mill.api.{Loose, PathRef, internal}
-import mill.buildfile.{AmmoniteParser, ParsedMillSetup, ReadDirectives}
+import mill.api.{Loose, PathRef, experimental}
+import mill.buildfile.{AmmoniteParser}
 import mill.define._
 import mill.scalalib.api.CompilationResult
-import mill.scalalib.bsp.BspBuildTarget
-import mill.scalalib.buildfile.MillBuildModule.millDiscover
 import mill.scalalib.{Dep, DepSyntax, Lib, ScalaModule, UnresolvedPath}
-import mill.{Agg, BuildInfo, T}
-import os.Path
-import upickle.default.{ReadWriter, macroRW}
+import mill.{Agg, T}
 
-/** A module that scans a mill project by inspecting the `build.sc`. */
-trait MillSetupScannerModule extends Module {
-
-//  object Deps {
-//    // incorrectly released without patch-level-scala-version
-//    val millModuledefs = ivy"com.lihaoyi::mill-main-moduledefs:${BuildInfo.millVersion}"
-//  }
-
-  /** The root path of the Mill project to scan. */
-  def millProjectPath: T[os.Path] = T { millSourcePath }
-
-  /** The Mill buildfile. */
-  def buildScFile: Source = T.source(millProjectPath() / "build.sc")
-
-  def supportUsingDirectives: T[Boolean] = false
-
-  /** The Mill project setup. */
-  def parsedMillSetup: T[ParsedMillSetup] = T {
-    val buildSc = buildScFile().path
-
-    val pass1 =
-      if (supportUsingDirectives()) {
-        ReadDirectives.readUsingDirectives(buildSc, parseVersionFiles = true)
-      } else {
-        ParsedMillSetup(
-          projectDir = millProjectPath(),
-          directives = ReadDirectives.readVersionFiles(buildSc / os.up),
-          buildScript = Option(buildScFile().path).filter(os.exists)
-        )
-      }
-    val pass2 = AmmoniteParser.parseAmmoniteImports(pass1)
-    pass2
-  }
-
-  def millVersion: T[String] = T {
-    parsedMillSetup().millVersion.getOrElse(BuildInfo.millVersion)
-  }
-
-  def millBinPlatform: T[String] = T {
-    millVersion().split("[.]") match {
-      case Array("0", "0" | "1" | "2" | "3" | "4" | "5", _*) => ""
-      case Array("0", a, _*) => s"0.${a}"
-      // TODO: support Milestone releases
-    }
-  }
-
-  def includedSourceFiles: T[Seq[PathRef]] = T {
-    parsedMillSetup().includedSourceFiles.map(PathRef(_))
-  }
-
-  protected def parseDeps: Task[String => Dep] = T.task { signature: String =>
-    val mv = millVersion()
-    val mbp = millBinPlatform()
-
-    // a missing version means, we want the Mill version
-    val millSig = signature.split("[:]") match {
-      case Array(org, "", pname, "", version)
-          if org.length > 0 && pname.length > 0 && version.length > 0 =>
-        s"${org}::${pname}_mill$$MILL_BIN_PLATFORM:${version}"
-      case Array(org, "", "", pname, "", version)
-          if org.length > 0 && pname.length > 0 && version.length > 0 =>
-        s"${org}:::${pname}_mill$$MILL_BIN_PLATFORM:${version}"
-      case Array(org, "", name) if org.length > 0 && name.length > 0 && signature.endsWith(":") =>
-        s"${org}::${name}:$$MILL_VERSION"
-      case _ => signature
-    }
-
-    // replace variables
-    val replaced = millSig
-      .replace("$MILL_VERSION", mv)
-      .replace("${MILL_VERSION}", mv)
-      .replace("$MILL_BIN_PLATFORM", mbp)
-      .replace("${MILL_BIN_PLATFORM}", mbp)
-
-    ivy"${replaced}"
-  }
-
-}
-
+@experimental
 trait MillBuildModule extends ScalaModule with MillSetupScannerModule {
 
   override def scalaVersion: T[String] = T {
@@ -106,17 +23,6 @@ trait MillBuildModule extends ScalaModule with MillSetupScannerModule {
 
   override def artifactSuffix: T[String] = T(s"${platformSuffix()}_${artifactScalaVersion()}")
 
-//  protected def millEmbeddedDeps: Input[Seq[String]] = T.input {
-//    millBinPlatform() match {
-//      case "" =>
-//    }
-////    Seq("com.lihaoyi::mill-scalalib_2.13:0.10.7-24-b7869c-DIRTY158cb392",
-  // "com.lihaoyi:mill-scalajslib_2.13:0.10.7-24-b7869c-DIRTY158cb392",
-  // "com.lihaoyi:mill-scalanativelib_2.13:0.10.7-24-b7869c-DIRTY158cb392",
-  // "com.lihaoyi:mill-bsp_2.13:0.10.7-24-b7869c-DIRTY158cb392")
-//    BuildInfo.millEmbeddedDeps
-//  }
-
   override def compileIvyDeps: T[Agg[Dep]] = T {
 //    Agg.from(millEmbeddedDeps().map(d => parseDeps(d)))
     val mv = millVersion()
@@ -125,7 +31,7 @@ trait MillBuildModule extends ScalaModule with MillSetupScannerModule {
       ivy"com.lihaoyi::mill-scalalib:${mv}",
       ivy"com.lihaoyi::mill-scalajslib:${mv}",
       ivy"com.lihaoyi::mill-scalanativelib:${mv}",
-      ivy"com.lihaoyi::mill-bsp:${mv}",
+      ivy"com.lihaoyi::mill-bsp:${mv}"
     )
   }
 
@@ -296,18 +202,7 @@ trait MillBuildModule extends ScalaModule with MillSetupScannerModule {
 //  )
 }
 
+@experimental
 object MillBuildModule extends ExternalModule with MillBuildModule {
   lazy val millDiscover: Discover[this.type] = Discover[this.type]
-}
-
-object MillNextBuildModule extends ExternalModule with MillBuildModule {
-  lazy val millDiscover: Discover[this.type] = Discover[this.type]
-
-  override def supportUsingDirectives: T[Boolean] = T(true)
-}
-
-case class WrappedSource(orig: PathRef, wrapped: Option[PathRef])
-
-object WrappedSource {
-  implicit val upickleRW: ReadWriter[WrappedSource] = macroRW
 }
