+++ scalalib/src/mill/scalalib/buildfile/WrappedSource.scala
@@ -1,0 +1,12 @@
+package mill.scalalib.buildfile
+
+import mill.api.{PathRef, experimental}
+import upickle.default.{ReadWriter, macroRW}
+
+@experimental
+case class WrappedSource(orig: PathRef, wrapped: Option[PathRef])
+
+@experimental
+object WrappedSource {
+  implicit val upickleRW: ReadWriter[WrappedSource] = macroRW
+}
